requirements.txt
python-dotenv==1.0.1
langchain-openai==0.0.5
langchain==0.1.6
